Restore working application from backup - complete React app with all features
- Fixed multiple default exports error
- Restored complete state management system
- User authentication & session management working
- Real-time task syncing with Firebase
- Handover notes system functional
- Wake-up calls management restored
- Breakfast time editing working
- Progress tracking and UI components all functional
- All modals and user interactions working properly

This version is stable and ready for production use.
